Change stout JSON::Value default to be JSON::Null.
The first type specified for the variant which is JSON::Value is the
default. Currently that is JSON::String which in turn defaults to
empty string. JSON::Null would be a more correct default value.

This change moves JSON::Null to the front of the declaration so that
it becomes the default and adds a unit test.

Review: https://reviews.apache.org/r/17568