Added support for launching tasks by TaskInfo.
This patch delegates the choice of executor to the containerizer by removing executorInfo dependencies up until Containerizer::launch().
Containerizer::launch() now returns a future to the executor info that is being run and the slave creates the corresponding executor structure when launch completes.
This means message handling from the running executor to the slave in the interim where the executor structure has not created, need to be enqueued until executor is ready.

Review: https://reviews.apache.org/r/18403