Introducing ContainerInfo as part of CommandInfo
Introduces the ContainerInfo protobuf as part of CommandInfo.
Right now, if present, the mesos containerizer fails the task launch to point out that we do not support it on that containerizer.

This will be needed for the ExternalContainerizer and possibly other containerizers as well.

Review: https://reviews.apache.org/r/20080