Changed executor state recovery to allow run recovery in absence of executor info.
This patch let executor recovery recover runs in the absence of
executor info.  This is needed as new task-info patch will introduce
an intermediate state where the executor info hasn't been check
pointed. In this interim, the slave may fail-over and should be in a
position to clean up orphan containers (as for now, the containerizer
API doesn't provide a way to reconcile the executor info and it is
therefore not possible to recover the containers in this case).

Review: https://reviews.apache.org/r/20221