Updated the Registrar to abort permanently upon encountering a Failure.
It's possible for a backed-up master (many items in its queue) to have
many operations enqueued in the Registrar.

In this event, the Master won't commit suicide until the initial
failure is processed. However, in the interim, subsequent operations
are potentially being performed against the Registrar. This could lead
to fighting between masters if a "demoted" master re-attempts to
acquirelog-leadership! This scenario can occur if the "demoted" master
has a large queue and the demotion event is towards the back of the
master's queue.

It would be preferable to ensure that after losing log leadership,
the "demoted" master does not try to re-acquire log leadership and
write to the log.

Review: https://reviews.apache.org/r/20981