Added --override flag to command executor.
It will be a common use case that external containerizer programs will
have to be able to control the command being carried out when the only
the task.command or container image is present.  The command executor
will in its current state try to run the command in the context of the
slave and not the container.

So, you could imagine the command executor being used by: 1) Replace
the command to run when the slave sends a runTask message 2) Prefix a
command to wrap call in when the slave sends a runTask message.

1) have been implemented and tested in this patch, alongside making
the command-executor "cli-ized" with help messages.

Review: https://reviews.apache.org/r/21329