Reduced acknowledgment delay in the scheduler driver.
We currently 'defer' in order to acknowledge processed status
updates in the scheduler driver.

This is unfortunate because it adds unnecessary delay and
unnecessary retries, especially when the scheduler's queue
is backed up.

It is sufficient to use 'aborted' to see if we can send a status
update. Ideally there would be proper locking on this variable
rather than the use of 'volatile', but since we're mostly concerned
about the driver aborting on the same thread as the status update
was processed, I've left it as is.

Review: https://reviews.apache.org/r/21788