Added workaround to allow the ExternalContainerizer containers to get recovered via slave state.
The slave currently expects a forked executor PID within its states when recovering. For the ExternalContainerizer however this does not make sense but is introduced in this patch to get around the problem as described in MESOS-1328 and MESOS-923. This patch can entirely be dropped if a fix for the mentioned tickets comes in first.

Review: https://reviews.apache.org/r/21677