Fix for web UI redirect message not showing up.
There was a bug that caused the "We're redirecting you to the leader"
message to not show up when visiting the web UI on a follower. Looks
to me like it was caused by an upgrade in twitter bootstrap, either
way it's a simple fix. I also changed the style from error to warning,
since the red message seemed a little harsh.

Review: https://reviews.apache.org/r/21850