Libevent: SSL: Changed ownership semantics of ssl connect socket.
libprocess Socket shares the ownership of the file descriptor with
libevent. In the destructor of the libprocess libevent_ssl socket, we
call ssl shutdown which is executed asynchronously. This causes the
libprocess socket file descriptor to be closed (and possibly reused)
when the same file descriptor could be used by libevent/ssl. Since we
set the shutdown options as SSL_RECEIVED_SHUTDOWN, we allow write
operations to continue with possibly closed file descriptor.

This change solves the above issue by copying(dup) the original file
descriptor and hands over the copy to libevent ssl. The copied
descriptor is then managed by libprocess Socket.

Review: https://reviews.apache.org/r/41253/