Simplified the rootfs preparation logic for command tasks.
This patch simplifies the way we deal with rootfs for command tasks.
Here are the major simplifications:

1) We no longer do a bind mount (<rootfs> -> <sandbox>/.rootfs) for
command tasks. All isolators can now assume that for both custom
executors or command tasks, if they change root filesystems, the root
filesystems will be specified in 'containerConfig.rootfs'. This will
simplify isolators' logic.

2) The sandbox bind mount will be done in the container's mount
namespace. Currently, it's done in the host mount namespace. This
creates a lot of complications, especially during the cleanup and
recovery path.

3) All persistent volumes will be consistently mounted under the
container's sandbox directory. Currently, depending on whether the
container changes the root filesystem or not, the persistent volumes
will be mounted at different locations. This simplifies the cleanup and
recovery path as well.

*** Modified for 0.28.2 ***

Review: https://reviews.apache.org/r/46807