Simplified `DRFSorter` to not track per-agent total resources.
The DRFSorter previously kept the total resources at each agent, along
with the total quantity of resources in the cluster. The latter figure
is what most of the clients of the sorter care about. In the few places
where the previous coding was using the per-agent total resources, it is
relatively easy to adjust the code to remove this dependency.

As part of this change, remove `total()` and
`update(const SlaveID& slaveId, const Resources& resources)` from the
Sorter interface. The former was unused; for the latter, code that used
it can instead be rewritten to adjust the total resources in the cluster
by first removing the previous resources at a agent and then adding
the new resources.

Review: https://reviews.apache.org/r/49375/